Rename :includes to :selectable
Also made a big change where field selector doesn't care if the
returned data is nested or not.

    property :foo, selectable: true

means :foo is not selected by default, and should explicitly be
opted-in with `?fields=foo`.

  class Comment
    property :commenter
  end

will render `post.commenter`, even when the commenter is a relation
and could return another nested data structure (in this case, a User
representation).

Also refactored so that Collection is now a proper subclass of
Property (actually, this can be further refactored so that we can
merge property and collection).