Support OAuth 2 code flow.
This required a couple of changes: a) ability to specify separate API
server in garage configuration (remote_server=) b) include
current_user as a helper in there as well, without inheriting from the
main ApplicationController.

a. was only necessary when your development environment is a
single-process server (e.g. Webrick) because otherwise your OAuth
client and server runs in the same process and exchanging OAuth code
will make a dead lock. In such case, run the API server as a different
process/port and specify that in the configuration, like spec/dummy
app does.