Allow procs to pass in selectable argument.
I experimented with returning 400 Bad Request so that scopes without
sufficient privilege can't just load the fields but also receive 400
error message saying it's invalid.

Facebook's `fields=` loader works that way, but that will break a lot
of existing clients requesting fields with wildcard.

For now, if a field that is not allowed is requested, it will silently
be omitted in the data.