Apply new gem name: the_garage
Another "garage" gem has already been hosted in RubyGems.org.
So, we've been suggested Bundler's github option to install this garage
gem, but it has some downsides. The biggest one is users can not install
the gem in the default way. Another one is users can not install the
gem while GitHub is down.